update c2wasm api
