+++ c2wasm-api/README.md
@@ -1,6 +1,6 @@
-# API that compiles C code to WASM
+# API that compiles C / JS code to WASM
 
-This directory contains a webserver for a C to WASM compiler and
+This directory contains a webserver for a C / JS to WASM compiler and
 Language Server (i.e. interactive linting).  Server is built with
 [Fastify](https://www.fastify.io/), fast and low overhead framework
 for Node.js
@@ -18,7 +18,7 @@ If you want to try this project follow these steps:
 - Install native dependencies (wasi-sdk & clangd) as automated by the docker container construction in the parent directory and run `yarn dev` here; alternatively, just run the docker instance from ../docker
 
 This should start server at port `:9000`, the actual compiling endpoint is this:
-[http://localhost:9000/api/build](localhost:9000/api/build)
+[http://localhost:9000/api/buildc](localhost:9000/api/buildc)
 
 Endpoint only accepts `HTTP POST`.
